pyproject.toml
[tool.poetry]
name = "discord-sales-bot"
version = "0.1.0"
description = ""
authors = ["RKL <hello@rl.games>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "1.7.3"
requests = "^2.27.1"
DateTime = "^4.4"
tweepy = "^4.6.0"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
mypy = "^0.931"
types-requests = "^2.27.11"
pre-commit = "^2.17.0"
python-dotenv = "^0.19.2"
ipykernel = "^6.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
